Computing area weighted covariates using two polygon sf or SpatVector objects

When poly_in and poly_weight are different classes, poly_weight will be converted to the class of poly_in.

Usage

aw_covariates(poly_in, poly_weight, id_poly_in = "ID")

Arguments

poly_in

A sf/SpatVector object at weighted means will be calculated.

poly_weight

A sf/SpatVector object from which weighted means will be calculated.

id_poly_in

character(1). The unique identifier of each polygon in poly_in

Value

A data.frame with all numeric fields of area-weighted means.

Author

Insang Song geoissong@gmail.com

Examples

# package
library(sf)
#> Linking to GEOS 3.10.2, GDAL 3.4.1, PROJ 8.2.1; sf_use_s2() is TRUE

# run
nc <- sf::st_read(system.file("shape/nc.shp", package="sf"))
#> Reading layer `nc' from data source 
#>   `/home/runner/work/_temp/Library/sf/shape/nc.shp' using driver `ESRI Shapefile'
#> Simple feature collection with 100 features and 14 fields
#> Geometry type: MULTIPOLYGON
#> Dimension:     XY
#> Bounding box:  xmin: -84.32385 ymin: 33.88199 xmax: -75.45698 ymax: 36.58965
#> Geodetic CRS:  NAD27
nc <- sf::st_transform(nc, 5070)
pp <- sf::st_sample(nc, size = 300)
pp <- sf::st_as_sf(pp)
pp[["id"]] <- seq(1, nrow(pp))
sf::st_crs(pp) <- "EPSG:5070"
ppb <- sf::st_buffer(pp, nQuadSegs=180, dist = units::set_units(20, 'km'))

system.time({ppb_nc_aw <- aw_covariates(ppb, nc, 'id')})
#>    user  system elapsed 
#>   0.350   0.008   0.358 
summary(ppb_nc_aw)
#>       AREA           PERIMETER         CNTY_         CNTY_ID    
#>  Min.   :0.05832   Min.   :1.117   Min.   :1825   Min.   :1825  
#>  1st Qu.:0.11831   1st Qu.:1.521   1st Qu.:1909   1st Qu.:1909  
#>  Median :0.13572   Median :1.707   Median :1999   Median :1999  
#>  Mean   :0.14448   Mean   :1.795   Mean   :2000   Mean   :2000  
#>  3rd Qu.:0.16716   3rd Qu.:1.979   3rd Qu.:2081   3rd Qu.:2081  
#>  Max.   :0.24098   Max.   :3.640   Max.   :2241   Max.   :2241  
#>      FIPSNO         CRESS_ID         BIR74             SID74       
#>  Min.   :37009   Min.   : 4.95   Min.   :  327.6   Min.   : 0.000  
#>  1st Qu.:37066   1st Qu.:33.70   1st Qu.: 1759.6   1st Qu.: 3.536  
#>  Median :37101   Median :50.77   Median : 2874.0   Median : 6.149  
#>  Mean   :37096   Mean   :48.62   Mean   : 3638.0   Mean   : 7.718  
#>  3rd Qu.:37124   3rd Qu.:62.51   3rd Qu.: 4294.2   3rd Qu.:10.230  
#>  Max.   :37185   Max.   :93.12   Max.   :18633.8   Max.   :37.249  
#>     NWBIR74             BIR79             SID79           NWBIR79      
#>  Min.   :   3.483   Min.   :  417.5   Min.   : 0.000   Min.   :   3.0  
#>  1st Qu.: 426.540   1st Qu.: 2134.0   1st Qu.: 4.190   1st Qu.: 558.4  
#>  Median : 938.207   Median : 3573.8   Median : 7.945   Median :1169.7  
#>  Mean   :1199.588   Mean   : 4672.2   Mean   : 9.152   Mean   :1550.4  
#>  3rd Qu.:1655.214   3rd Qu.: 5367.0   3rd Qu.:12.389   3rd Qu.:2080.6  
#>  Max.   :6846.306   Max.   :26502.3   Max.   :52.022   Max.   :9900.0  
#>           geometry  
#>  POLYGON      :300  
#>  epsg:5070    :  0  
#>  +proj=aea ...:  0  
#>                     
#>                     
#>                     
#### Example of aw_covariates ends ####