Calculate SEDC covariates

Calculate SEDC covariates

Usage

calculate_sedc(
  point_from,
  point_to,
  id,
  sedc_bandwidth,
  threshold,
  target_fields
)

Arguments

point_from

SpatVector object. Locations where the sum of SEDCs are calculated.

point_to

SpatVector object. Locations where each SEDC is calculated.

id

character(1). Name of the unique id field in point_to.

sedc_bandwidth

numeric(1). Distance at which the source concentration is reduced to exp(-3) (approximately -95 %)

threshold

numeric(1). For computational efficiency, the nearest points in threshold will be selected. Default is 2 * sedc_bandwidth.

target_fields

character(varying). Field names in characters.

Value

data.frame (tibble) object with input field names with a suffix "_sedc" where the sums of EDC are stored. Additional attributes are attached for the EDC information.

attr(result, "sedc_bandwidth"): the bandwidth where concentration reduces to approximately five percent

attr(result, "sedc_threshold"): the threshold distance at which emission source points are excluded beyond that

Note

Distance calculation is done with terra functions internally. Thus, the function internally converts sf objects in point_* arguments to terra. The optimal EDC should be carefully chosen by users.

Author

Insang Song