Check the subject object and perform necessary conversions if needed.

This function checks the class of the input object and performs necessary conversions if needed.

Usage

check_subject(subject, extent = NULL, subject_id = NULL)

Arguments

subject

The input object to be checked.

extent

The extent of the subject object.

subject_id

Optional. The ID of the subject object.

Value

The checked and converted subject object.

Examples

# Check a SpatVector object
ncpath <- system.file("gpkg/nc.gpkg", package = "sf")
ncsf <- sf::st_read(ncpath)
#> Reading layer `nc.gpkg' from data source 
#>   `/home/runner/work/_temp/Library/sf/gpkg/nc.gpkg' using driver `GPKG'
#> Simple feature collection with 100 features and 14 fields
#> Geometry type: MULTIPOLYGON
#> Dimension:     XY
#> Bounding box:  xmin: -84.32385 ymin: 33.88199 xmax: -75.45698 ymax: 36.58965
#> Geodetic CRS:  NAD27
extent <- c(-80, -77, 35, 36)
check_subject(subject = ncsf, extent = extent, subject_id = "FIPS")
#>  class       : SpatVector 
#>  geometry    : polygons 
#>  dimensions  : 100, 14  (geometries, attributes)
#>  extent      : -84.32385, -75.45698, 33.88199, 36.58965  (xmin, xmax, ymin, ymax)
#>  coord. ref. : lon/lat NAD27 (EPSG:4267) 
#>  names       :  AREA PERIMETER CNTY_ CNTY_ID      NAME  FIPS    FIPSNO CRESS_ID
#>  type        : <num>     <num> <num>   <num>     <chr> <chr>     <num>    <int>
#>  values      : 0.114     1.442  1825    1825      Ashe 37009 3.701e+04        5
#>                0.061     1.231  1827    1827 Alleghany 37005   3.7e+04        3
#>                0.143      1.63  1828    1828     Surry 37171 3.717e+04       86
#>  BIR74 SID74 (and 4 more)
#>  <num> <num>             
#>   1091     1             
#>    487     0             
#>   3188     5             

# Check a character object
check_subject(subject = ncpath, extent = extent, subject_id = "FIPS")
#>  class       : SpatVector 
#>  geometry    : polygons 
#>  dimensions  : 33, 14  (geometries, attributes)
#>  extent      : -80.32528, -76.49254, 34.55375, 36.50989  (xmin, xmax, ymin, ymax)
#>  source      : nc.gpkg (nc.gpkg)
#>  coord. ref. : lon/lat NAD27 (EPSG:4267) 
#>  names       :  AREA PERIMETER CNTY_ CNTY_ID       NAME  FIPS    FIPSNO
#>  type        : <num>     <num> <num>   <num>      <chr> <chr>     <num>
#>  values      :  0.13     1.732  2044    2044 Montgomery 37123 3.712e+04
#>                0.201     1.805  1968    1968   Randolph 37151 3.715e+04
#>                 0.17      1.68  1903    1903   Guilford 37081 3.708e+04
#>  CRESS_ID     BIR74 SID74 (and 4 more)
#>     <int>     <num> <num>             
#>        62      1258     3             
#>        76      4456     7             
#>        41 1.618e+04    23