Process a given function in the entire or partial computational grids

Currently only accepting multicore setting (single node, single process, and multiple threads). For details of the terminology in future package, refer to plan. This function assumes that users have one raster file and a sizable and spatially distributed target locations. Each thread will process ceiling(|Ng|/|Nt|) grids where |Ng| denotes the number of grids and |Nt| denotes the number of threads.

Usage

distribute_process_grid(
  grids,
  grid_target_id = NULL,
  debug = FALSE,
  fun_dist,
  ...
)

Arguments

grids

sf/SpatVector object. Computational grids. It takes a strict assumption that the grid input is an output of get_computational_regions

grid_target_id

character(1) or numeric(2). Default is NULL. If NULL, all grid_ids are used. "id_from:id_to" format or c(unique(grid_id)[id_from], unique(grid_id)[id_to])

debug

logical(1). Prints error messages if there were any errors during the calculation.

fun_dist

sf, terra or chopin functions.

...

Arguments passed to the argument fun_dist.

Value

a data.frame object with computation results. For entries of the results, consult the function used in fun_dist argument.

Note

In dynamic dots (...), the first and second arguments should be the fun_dist arguments where sf/SpatVector objects are accepted. Virtually any sf/terra functions that accept two arguments can be put in fun_dist, but please be advised that some spatial operations do not necessarily give the exact result from what would have been done single-thread. For example, distance calculated through this function may return the lower value than actual because the computational region was reduced. This would be the case especially where the target features are spatially sparsely distributed.

Author

Insang Song geoissong@gmail.com

Examples

library(future)
plan(multicore, workers = 4)
# Does not run ...
# distribute_process_grid()