Extract summarized values from raster with points and a buffer radius

For simplicity, it is assumed that the coordinate systems of the points and the raster are the same.

Usage

extract_at_buffer(
  points = NULL,
  surf = NULL,
  radius = NULL,
  id = NULL,
  qsegs = 90L,
  func = "mean",
  kernel = NULL,
  bandwidth = NULL
)

extract_at_buffer_flat(
  points = NULL,
  surf = NULL,
  radius = NULL,
  id = NULL,
  qsegs = NULL,
  func = "mean",
  kernel = NULL,
  bandwidth = NULL
)

extract_at_buffer_kernel(
  points = NULL,
  surf = NULL,
  radius = NULL,
  id = NULL,
  qsegs = NULL,
  func = stats::weighted.mean,
  kernel = NULL,
  bandwidth = NULL
)

Arguments

points

sf/SpatVector object. Coordinates where buffers will be generated.

surf

SpatRaster object. A raster at which summary will be calculated

radius

numeric(1). Buffer radius. here we assume circular buffers only

id

character(1). Unique identifier of each point.

qsegs

integer(1). Number of vertices at a quarter of a circle. Default is 90L.

func

a function taking a numeric vector argument.

kernel

character(1). Name of a kernel function One of "uniform", "triweight", "quartic", and "epanechnikov"

bandwidth

numeric(1). Kernel bandwidth.

Value

a data.frame object with mean value

Author

Insang Song geoissong@gmail.com

Examples

library(terra)
rrast <- terra::rast(nrow = 100, ncol = 100)
terra::crs(rrast) <- "EPSG:5070"
terra::values(rrast) <- rgamma(1e4, 4, 2)
rpnt <- terra::spatSample(rrast, 100L, as.points = TRUE)
rpnt$pid <- sprintf("id_%03d", seq(1, 100))
extract_at_buffer(rpnt, rrast, 4, "pid")
#>        pid     mean
#> 1   id_001 2.088129
#> 2   id_002 1.801404
#> 3   id_003 1.818624
#> 4   id_004 1.713521
#> 5   id_005 1.861658
#> 6   id_006 1.719859
#> 7   id_007 1.974695
#> 8   id_008 2.049741
#> 9   id_009 2.316861
#> 10  id_010 1.916453
#> 11  id_011 2.355014
#> 12  id_012 2.299939
#> 13  id_013 2.039365
#> 14  id_014 2.103469
#> 15  id_015 2.284654
#> 16  id_016 2.215988
#> 17  id_017 1.960786
#> 18  id_018 3.082453
#> 19  id_019 2.374119
#> 20  id_020 2.151488
#> 21  id_021 1.687319
#> 22  id_022 1.919163
#> 23  id_023 2.392955
#> 24  id_024 1.938812
#> 25  id_025 1.650142
#> 26  id_026 1.787531
#> 27  id_027 2.072593
#> 28  id_028 2.150475
#> 29  id_029 1.506016
#> 30  id_030 1.740266
#> 31  id_031 1.698756
#> 32  id_032 2.337476
#> 33  id_033 2.223276
#> 34  id_034 1.862528
#> 35  id_035 2.522879
#> 36  id_036 1.633430
#> 37  id_037 2.207080
#> 38  id_038 1.562899
#> 39  id_039 1.925542
#> 40  id_040 2.149704
#> 41  id_041 2.031895
#> 42  id_042 2.096056
#> 43  id_043 2.927146
#> 44  id_044 1.958425
#> 45  id_045 1.724679
#> 46  id_046 2.121109
#> 47  id_047 1.814018
#> 48  id_048 2.495163
#> 49  id_049 1.461759
#> 50  id_050 2.375083
#> 51  id_051 2.430701
#> 52  id_052 2.516381
#> 53  id_053 1.902838
#> 54  id_054 2.462683
#> 55  id_055 2.037783
#> 56  id_056 2.275699
#> 57  id_057 2.049863
#> 58  id_058 2.527668
#> 59  id_059 2.154519
#> 60  id_060 1.581526
#> 61  id_061 1.722006
#> 62  id_062 1.939144
#> 63  id_063 2.082714
#> 64  id_064 2.135763
#> 65  id_065 2.703117
#> 66  id_066 1.497881
#> 67  id_067 1.821361
#> 68  id_068 1.566550
#> 69  id_069 2.092789
#> 70  id_070 2.161509
#> 71  id_071 2.113995
#> 72  id_072 2.036025
#> 73  id_073 2.370155
#> 74  id_074 2.151327
#> 75  id_075 2.080855
#> 76  id_076 1.938411
#> 77  id_077 2.040131
#> 78  id_078 1.703379
#> 79  id_079 2.363913
#> 80  id_080 1.449415
#> 81  id_081 1.581289
#> 82  id_082 2.187207
#> 83  id_083 1.777520
#> 84  id_084 1.634075
#> 85  id_085 2.001484
#> 86  id_086 1.976651
#> 87  id_087 1.747148
#> 88  id_088 1.643719
#> 89  id_089 1.863137
#> 90  id_090 2.511853
#> 91  id_091 2.203949
#> 92  id_092 2.331332
#> 93  id_093 1.458342
#> 94  id_094 1.926339
#> 95  id_095 1.396570
#> 96  id_096 1.898611
#> 97  id_097 2.262470
#> 98  id_098 2.511716
#> 99  id_099 2.503972
#> 100 id_100 1.715506