Generate grid polygons

Returns a sf object that includes x- and y- index by using two inputs ncutsx and ncutsy, which are x- and y-directional splits, respectively.

Usage

par_make_grid(points_in = NULL, ncutsx = NULL, ncutsy = NULL)

Arguments

points_in

sf or SpatVector object. Target points of computation. character(1) of file path is also acceptable.

ncutsx

integer(1). The number of splits along x-axis.

ncutsy

integer(1). The number of splits along y-axis.

Value

A sf or SpatVector object of computation grids with unique grid id (CGRIDID).

Note

Grids are generated based on the extent of points_in first, then exhaustive grids will be filtered by the intersection between these and points_in. Thus, the number of generated grids may be smaller than ncutsx * ncutsy.

See also

Author

Insang Song

Examples

if (FALSE) { # \dontrun{
library(sf)
library(terra)
options(sf_use_s2 = FALSE)

nc_path <- system.file("gpkg/nc.gpkg", package = "sf")
nc <- terra::vect(nc_path)
nc_rp <- terra::spatSample(nc, 1000)
nc_gr <- par_make_grid(nc_rp, 10L, 6L)

plot(nc_rp)
plot(nc_gr, add = TRUE)
} # }