Map arguments to the desired names

This function maps the arguments of a target function to the desired names. Users will use a named list name_match to standardize the argument names, at least x and y, to the target function. This function is particularly useful to parallelize functions for spatial data outside sf and terra packages that do not have arguments named x and/or y. par_* functions could detect such functions by wrapping nonstandardized functions to parallelize the computation.

Usage

par_map_args(fun, name_match = list(), ...)

Arguments

fun

The target function to be called.

name_match

A named list of arguments to be mapped.

...

Arguments to be passed to the target function.

Value

The result of the target function.

See also

Examples

# Example target function
example_fun <- function(x, y, z = 1) {
  return(c(x = x, y = y, z = z))
}

# Example usage of map_args_xy
result <- par_map_args(fun = example_fun,
                       name_match = list(a = "x", b = "y"),
                       a = 10, b = 20, z = 5)
print(result)
#>  x  y  z 
#> 10 20  5