Area weighted summary using two polygon sf or SpatVector objects

When poly_in and poly_weight are different classes, poly_weight will be converted to the class of poly_in.

Usage

summarize_aw(
  poly_in = NULL,
  poly_weight = NULL,
  target_fields = NULL,
  id_poly_in = "ID",
  fun = stats::weighted.mean
)

Arguments

poly_in

A sf/SpatVector object at weighted means will be calculated.

poly_weight

A sf/SpatVector object from which weighted means will be calculated.

target_fields

character. Field names to calculate area-weighted .

id_poly_in

character(1). The unique identifier of each polygon in poly_in. Default is "ID".

fun

function(1). The function to calculate the weighted summary. Default is stats::weighted.mean. The function must have a w argument.

Value

A data.frame with all numeric fields of area-weighted means.

See also

Author

Insang Song geoissong@gmail.com

Examples

# package
library(sf)
sf_use_s2(FALSE)
nc <- sf::st_read(system.file("shape/nc.shp", package="sf"))
#> Reading layer `nc' from data source 
#>   `/home/runner/work/_temp/Library/sf/shape/nc.shp' using driver `ESRI Shapefile'
#> Simple feature collection with 100 features and 14 fields
#> Geometry type: MULTIPOLYGON
#> Dimension:     XY
#> Bounding box:  xmin: -84.32385 ymin: 33.88199 xmax: -75.45698 ymax: 36.58965
#> Geodetic CRS:  NAD27
nc <- sf::st_transform(nc, 5070)
pp <- sf::st_sample(nc, size = 300)
pp <- sf::st_as_sf(pp)
pp[["id"]] <- seq(1, nrow(pp))
sf::st_crs(pp) <- "EPSG:5070"
ppb <- sf::st_buffer(pp, nQuadSegs=180, dist = units::set_units(20, "km"))

system.time(ppb_nc_aw <- summarize_aw(ppb, nc, c("BIR74", "BIR79"), "id"))
#>    user  system elapsed 
#>   0.344   0.000   0.344 
summary(ppb_nc_aw)
#>      BIR74             BIR79                  geometry  
#>  Min.   :  265.1   Min.   :  348.3   POLYGON      :300  
#>  1st Qu.: 1932.8   1st Qu.: 2422.2   epsg:5070    :  0  
#>  Median : 3116.3   Median : 3855.7   +proj=aea ...:  0  
#>  Mean   : 4006.6   Mean   : 5166.8                      
#>  3rd Qu.: 4855.9   3rd Qu.: 6216.7                      
#>  Max.   :16721.9   Max.   :23772.3                      
#### Example of summarize_aw ends ####